+++ src/main/java/com/server/pitch/PitchApplication.java
@@ -2,19 +2,12 @@ package com.server.pitch;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.context.annotation.Bean;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 
 @SpringBootApplication
 public class PitchApplication {
 
 	public static void main(String[] args) {
 		SpringApplication.run(PitchApplication.class, args);
-	}
-
-	@Bean
-	public BCryptPasswordEncoder passwordEncoder(){
-		return new BCryptPasswordEncoder();
 	}
 
 }
