+++ src/main/java/com/server/pitch/loginuser/controller/Controller.java
@@ -1,4 +1,4 @@
-package com.server.pitch.users.controller;
+package com.server.pitch.loginuser.controller;
 
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
