+++ src/main/java/com/server/pitch/loginuser/mapper/LoginUserMapper.java
@@ -1,0 +1,11 @@
+package com.server.pitch.loginuser.mapper;
+
+import com.server.pitch.loginuser.domain.LoginUser;
+import org.apache.ibatis.annotations.Mapper;
+
+import java.util.List;
+
+@Mapper
+public interface LoginUserMapper {
+    public List<LoginUser> list();
+}
