--- src/main/resources/application-oauth.yml
@@ -1,20 +1,0 @@
-#
-#spring:
-#  security:
-#    oauth2:
-#      client:
-#        registration:
-#          google:
-#            client-id: 1061330767873-0hsk6i0platdln04io0velba0ojonsc0.apps.googleusercontent.com
-#            client-secret: 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
-#            redirect-uri: http://localhost:8888/auth/google
-#            token-uri: https://oauth2.googleapis.com/token
-#            auth-uri: https://accounts.google.com/o/oauth2/auth
-
-oauth2:
-  google:
-    client-id: 1061330767873-0hsk6i0platdln04io0velba0ojonsc0.apps.googleusercontent.com
-    client-secret: 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
-    redirect-uri: http://localhost:8888/auth/google
-    token-uri: https://oauth2.googleapis.com/token
-    auth-uri: https://www.googleapis.com/oauth2/v2/userinfo
