+++ src/main/resources/application.yml
@@ -7,26 +7,11 @@ spring:
     url: jdbc:mariadb://www.soyoung.dev:8690/pitch
     username: dev
     password: best2
-  redis:
-    port: 6379
-    host: www.soyoung.dev
-    password: 1323
-  profiles:
-    include: oauth
 
 mybatis:
   mapper-locations: classpath:mapper/*.xml
   type-aliases-package: com.server.pitch
 
-token:
-  expiration_time: 86400000
-  refreshToken_time: 604800000
-  asecret: jwtAccess
-  rsecret: jwtRefresh
-
-
-#logging:
-#  level:
-#    org.springframework.security: DEBUG
-#    com.server.pitch.filter: DEBUG
-
+logging:
+  level:
+    org.springframework.security: DEBUG
