+++ src/main/resources/mapper/LoginUserMapper.xml
@@ -1,0 +1,13 @@
+<?xml version="1.0" encoding="UTF-8" ?>
+<!DOCTYPE mapper
+        PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
+        "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
+
+<mapper namespace="org.server.pitch.loginuser.mapper.LoginUserMapper">
+<!--    테스트용 -->
+    <select id="list" resultType="LoginUser">
+        select user_id, user_pw, user_nm, role from loginuser
+    </select>
+
+
+</mapper>
