--- src/main/resources/mapper/UsersMapper.xml
@@ -1,29 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-<!DOCTYPE mapper
-        PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
-        "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-
-<mapper namespace="com.server.pitch.users.mapper.UsersMapper">
-    <select id="list" resultType="Users">
-        select user_id, user_pw, user_nm, role from users
-    </select>
-
-    <select id="selectUser" resultType="Users">
-        select user_id, user_email, user_pw, user_nm, role from users
-        <where>
-            <if test="user_id != null">
-                AND user_id = #{user_id}
-            </if>
-            <if test="user_email != null">
-                AND user_email = #{user_email}
-            </if>
-        </where>
-    </select>
-
-    <insert id="insertUser">
-        insert into users (user_id, user_email, user_pw, user_nm, user_phone, user_birth)
-        values (#{user_id}, #{user_email}, #{user_pw}, #{user_nm}, #{user_phone}, #{user_birth})
-    </insert>
-
-
-</mapper>
